Update README.md
In this link you'll find how I learned how to "Activate Directory Home Lab with PowerShell". This gave me the opportunity to have some level of hands-on experience, and this completely helped my understanding that has now deepened with a great amount of knowledge that I can apply within cybersecurity.